--- a/kclvm/config/src/settings.rs
+++ b/kclvm/config/src/settings.rs
@@ -403,6 +403,55 @@ pub fn merge_settings(settings: &[SettingsFile]) -> SettingsFile {
     result
 }
 
+/// Build SettingsPathBuf from args.
+pub fn build_settings_pathbuf(
+    files: &[&str],
+    setting_files: Option<Vec<&str>>,
+    setting_config: Option<SettingsFile>,
+) -> Result<SettingsPathBuf> {
+    let mut path = None;
+    let settings = if let Some(files) = setting_files {
+        let mut settings = vec![];
+        for file in &files {
+            let s = load_file(file)?;
+            if !s.input().is_empty() {
+                path = Some(
+                    PathBuf::from(file)
+                        .parent()
+                        .map(|p| p.to_path_buf())
+                        .ok_or(anyhow::anyhow!("The parent path of {file} is not found"))?,
+                )
+            }
+            settings.push(s);
+        }
+        merge_settings(&settings)
+    // If exists default kcl.yaml, load it.
+    } else if std::fs::metadata(DEFAULT_SETTING_FILE).is_ok() {
+        path = Some(
+            PathBuf::from(DEFAULT_SETTING_FILE)
+                .parent()
+                .map(|p| p.to_path_buf())
+                .ok_or(anyhow::anyhow!(
+                    "The parent path of {DEFAULT_SETTING_FILE} is not found"
+                ))?,
+        );
+        load_file(DEFAULT_SETTING_FILE)?
+    } else {
+        SettingsFile::default()
+    };
+    let mut settings = if let Some(setting_config) = setting_config {
+        merge_settings(&[settings, setting_config])
+    } else {
+        settings
+    };
+    if let Some(config) = &mut settings.kcl_cli_configs {
+        if !files.is_empty() {
+            config.files = Some(files.iter().map(|f| f.to_string()).collect());
+        }
+    }
+    Ok(SettingsPathBuf::new(path, settings))
+}
+
 #[cfg(test)]
 mod settings_test {
     use crate::settings::*;
@@ -465,53 +514,4 @@ mod settings_test {
         }
         Ok(())
     }
-}
-
-/// Build SettingsPathBuf from args.
-pub fn build_settings_pathbuf(
-    files: &[&str],
-    setting_files: Option<Vec<&str>>,
-    setting_config: Option<SettingsFile>,
-) -> Result<SettingsPathBuf> {
-    let mut path = None;
-    let settings = if let Some(files) = setting_files {
-        let mut settings = vec![];
-        for file in &files {
-            let s = load_file(file)?;
-            if !s.input().is_empty() {
-                path = Some(
-                    PathBuf::from(file)
-                        .parent()
-                        .map(|p| p.to_path_buf())
-                        .ok_or(anyhow::anyhow!("The parent path of {file} is not found"))?,
-                )
-            }
-            settings.push(s);
-        }
-        merge_settings(&settings)
-    // If exists default kcl.yaml, load it.
-    } else if std::fs::metadata(DEFAULT_SETTING_FILE).is_ok() {
-        path = Some(
-            PathBuf::from(DEFAULT_SETTING_FILE)
-                .parent()
-                .map(|p| p.to_path_buf())
-                .ok_or(anyhow::anyhow!(
-                    "The parent path of {DEFAULT_SETTING_FILE} is not found"
-                ))?,
-        );
-        load_file(DEFAULT_SETTING_FILE)?
-    } else {
-        SettingsFile::default()
-    };
-    let mut settings = if let Some(setting_config) = setting_config {
-        merge_settings(&[settings, setting_config])
-    } else {
-        settings
-    };
-    if let Some(config) = &mut settings.kcl_cli_configs {
-        if !files.is_empty() {
-            config.files = Some(files.iter().map(|f| f.to_string()).collect());
-        }
-    }
-    Ok(SettingsPathBuf::new(path, settings))
 }